refactor: extracted starter project code from create options
