+++ pkg/odo/cli/component/create.go
@@ -754,177 +754,6 @@ func (co *CreateOptions) Validate() (err error) {
 	return nil
 }
 
-// downloadGitProject downloads the git starter projects from devfile.yaml
-func downloadGitProject(starterProject *devfilev1.StarterProject, starterToken, path string) error {
-
-	var projectSource devfilev1.GitLikeProjectSource
-	if starterProject.Git != nil {
-		projectSource = starterProject.Git.GitLikeProjectSource
-	} else {
-		projectSource = starterProject.Github.GitLikeProjectSource
-	}
-
-	remoteName, remoteUrl, revision, err := parsercommon.GetDefaultSource(projectSource)
-	if err != nil {
-		return errors.Wrapf(err, "unable to get default project source for starter project %s", starterProject.Name)
-	}
-
-	// convert revision to referenceName type, ref name could be a branch or tag
-	// if revision is not specified it would be the default branch of the project
-	refName := plumbing.ReferenceName(revision)
-
-	if plumbing.IsHash(revision) {
-		// Specifying commit in the reference name is not supported by the go-git library
-		// while doing git.PlainClone()
-		log.Warning("Specifying commit in 'revision' is not yet supported in odo.")
-		// overriding revision to empty as we do not support this
-		revision = ""
-	}
-
-	if revision != "" {
-		// lets consider revision to be a branch name first
-		refName = plumbing.NewBranchReferenceName(revision)
-	}
-
-	downloadSpinner := log.Spinnerf("Downloading starter project %s from %s", starterProject.Name, remoteUrl)
-	defer downloadSpinner.End(false)
-
-	cloneOptions := &git.CloneOptions{
-		URL:           remoteUrl,
-		RemoteName:    remoteName,
-		ReferenceName: refName,
-		SingleBranch:  true,
-		// we don't need history for starter projects
-		Depth: 1,
-	}
-
-	if starterToken != "" {
-		cloneOptions.Auth = &http.BasicAuth{
-			Username: registryUtil.RegistryUser,
-			Password: starterToken,
-		}
-	}
-
-	originalPath := ""
-	if starterProject.SubDir != "" {
-		originalPath = path
-		path, err = ioutil.TempDir("", "")
-		if err != nil {
-			return err
-		}
-	}
-
-	_, err = git.PlainClone(path, false, cloneOptions)
-
-	if err != nil {
-
-		// it returns the following error if no matching ref found
-		// if we get this error, we are trying again considering revision as tag, only if revision is specified.
-		if _, ok := err.(git.NoMatchingRefSpecError); !ok || revision == "" {
-			return err
-		}
-
-		// try again to consider revision as tag name
-		cloneOptions.ReferenceName = plumbing.NewTagReferenceName(revision)
-		// remove if any .git folder downloaded in above try
-		_ = os.RemoveAll(filepath.Join(path, ".git"))
-		_, err = git.PlainClone(path, false, cloneOptions)
-		if err != nil {
-			return err
-		}
-	}
-
-	// we don't want to download project be a git repo
-	err = os.RemoveAll(filepath.Join(path, ".git"))
-	if err != nil {
-		// we don't need to return (fail) if this happens
-		log.Warning("Unable to delete .git from cloned starter project")
-	}
-
-	if starterProject.SubDir != "" {
-		err = util.GitSubDir(path, originalPath,
-			starterProject.SubDir)
-		if err != nil {
-			return err
-		}
-	}
-	downloadSpinner.End(true)
-
-	return nil
-
-}
-
-// Downloads first starter project from list of starter projects in devfile
-// Currently type git with a non github url is not supported
-func (co *CreateOptions) downloadStarterProject(devObj parser.DevfileObj, projectPassed string, interactive bool) error {
-	if projectPassed == "" && !interactive {
-		return nil
-	}
-
-	// Retrieve starter projects
-	starterProjects := devObj.Data.GetStarterProjects()
-
-	var starterProject *devfilev1.StarterProject
-	var err error
-	if interactive {
-		starterProject = getStarterProjectInteractiveMode(starterProjects)
-	} else {
-		starterProject, err = getStarterProjectFromFlag(starterProjects, projectPassed)
-		if err != nil {
-			return err
-		}
-	}
-
-	if starterProject == nil {
-		return nil
-	}
-
-	// Retrieve the working directory in order to clone correctly
-	path, err := os.Getwd()
-	if err != nil {
-		return errors.Wrapf(err, "Could not get the current working directory.")
-	}
-
-	// We will check to see if the project has a valid directory
-	err = util.IsValidProjectDir(path, DevfilePath)
-	if err != nil {
-		return err
-	}
-
-	if co.devfileMetadata.starterToken == "" && registryUtil.IsSecure(co.devfileMetadata.devfileRegistry.Name) {
-		token, err := keyring.Get(fmt.Sprintf("%s%s", util.CredentialPrefix, co.devfileMetadata.devfileRegistry.Name), registryUtil.RegistryUser)
-		if err != nil {
-			return errors.Wrap(err, "unable to get secure registry credential from keyring")
-		}
-		co.devfileMetadata.starterToken = token
-	}
-
-	log.Info("\nStarter Project")
-
-	if starterProject.Git != nil || starterProject.Github != nil {
-		err := downloadGitProject(starterProject, co.devfileMetadata.starterToken, path)
-
-		if err != nil {
-			return err
-		}
-
-	} else if starterProject.Zip != nil {
-		url := starterProject.Zip.Location
-		sparseDir := starterProject.SubDir
-		downloadSpinner := log.Spinnerf("Downloading starter project %s from %s", starterProject.Name, url)
-		err := co.checkoutProject(sparseDir, url, path)
-		if err != nil {
-			downloadSpinner.End(false)
-			return err
-		}
-		downloadSpinner.End(true)
-	} else {
-		return errors.Errorf("Project type not supported")
-	}
-
-	return nil
-}
-
 func (co *CreateOptions) s2iRun() (err error) {
 	err = co.LocalConfigInfo.SetComponentSettings(co.componentSettings)
 	if err != nil {
@@ -1101,87 +930,6 @@ func (co *CreateOptions) Run() (err error) {
 		machineoutput.OutputSuccess(componentDesc)
 	}
 	return
-}
-
-func (co *CreateOptions) checkoutProject(subDir, zipURL, path string) error {
-	if subDir != "" {
-		err := util.GetAndExtractZip(zipURL, path, subDir, co.devfileMetadata.starterToken)
-		if err != nil {
-			return errors.Wrap(err, "failed to download and extract project zip folder")
-		}
-	} else {
-		// extract project to current working directory
-		err := util.GetAndExtractZip(zipURL, path, "/", co.devfileMetadata.starterToken)
-		if err != nil {
-			return errors.Wrap(err, "failed to download and extract project zip folder")
-		}
-	}
-	return nil
-}
-
-// getStarterProjectFromFlag gets starter project value from flag --starter.
-func getStarterProjectFromFlag(projects []devfilev1.StarterProject, projectPassed string) (project *devfilev1.StarterProject, err error) {
-
-	nOfProjects := len(projects)
-
-	if nOfProjects == 0 {
-		return nil, errors.Errorf("no starter project found in devfile.")
-	}
-
-	// Determine what project to be used
-	if nOfProjects == 1 && projectPassed == defaultStarterProjectName {
-		project = &projects[0]
-	} else if nOfProjects > 1 && projectPassed == defaultStarterProjectName {
-		project = &projects[0]
-		log.Warning("There are multiple projects in this devfile but none have been specified in --starter. Downloading the first: " + project.Name)
-	} else { //If the user has specified a project
-		projectFound := false
-		for indexOfProject, projectInfo := range projects {
-			if projectInfo.Name == projectPassed { //Get the index
-				project = &projects[indexOfProject]
-				projectFound = true
-				break
-			}
-		}
-
-		if !projectFound {
-			return nil, errors.Errorf("the project: %s specified in --starter does not exist", projectPassed)
-		}
-	}
-
-	return project, err
-
-}
-
-func (co *CreateOptions) SetComponentSettings(args []string) error {
-	err := co.setComponentSourceAttributes()
-	if err != nil {
-		return err
-	}
-	err = co.setComponentName(args)
-	if err != nil {
-		return err
-	}
-
-	var portList []string
-	if len(co.componentPorts) > 0 {
-		portList = co.componentPorts
-	} else {
-		portList, err = co.Client.GetPortsFromBuilderImage(*co.componentSettings.Type)
-		if err != nil {
-			return err
-		}
-	}
-
-	co.componentSettings.Ports = &(portList)
-	co.componentSettings.Project = &(co.Context.Project)
-	envs, err := config.NewEnvVarListFromSlice(co.componentEnvVars)
-	if err != nil {
-		return err
-	}
-	co.componentSettings.Envs = envs
-	co.ignores = []string{}
-	return nil
 }
 
 // NewCmdCreate implements the create odo command
